rfidclient/requirements.txt
pyserial
pyautogui
pyperclip